10_batch_processing/tests/unit/test_batch_processing_stack.py
import aws_cdk as core
import aws_cdk.assertions as assertions

from batch_processing.batch_processing_stack import BatchProcessingStack

# example tests. To run these tests, uncomment this file along with the example
# resource in batch_processing/batch_processing_stack.py
def test_sqs_queue_created():
    app = core.App()
    stack = BatchProcessingStack(app, "batch-processing")
    template = assertions.Template.from_stack(stack)

#     template.has_resource_properties("AWS::SQS::Queue", {
#         "VisibilityTimeout": 300
#     })
